backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose
python-dotenv
groq
httpx
cryptography
pyjwt
